refactor eliminate.rs as module processor
	modified:   src/cdb/db.rs
	modified:   src/lib.rs
	new file:   src/processor/eliminate.rs
	new file:   src/processor/heap.rs
	renamed:    src/eliminate.rs -> src/processor/mod.rs
	new file:   src/processor/subsume.rs
	modified:   src/solver/build.rs
	modified:   src/solver/mod.rs
	modified:   src/solver/search.rs
	modified:   src/state.rs
